style: fix issues during review
