+++ src/components/directives/image/test/const.ts
@@ -10,6 +10,6 @@ export const BROKEN_PICTURE_SRC = 'this.is.broken.picture.png';
 export const SLOW_LOAD_PICTURE_SRC = 'this.is.slow-load.picture.png';
 
 /**
- * A red square 20x20 px in size, encoded as base64.
+ * A red square 20x20 pixels in size, encoded as base64.
  */
 export const EXISTING_PICTURE_SRC = 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IArs4c6QAAACtJREFUOE9jfCuj8p+BioBx1ECKQ3M0DCkOQobRMBwNQzJCYDTZkBFoaFoAMcorhYzB4yEAAAAASUVORK5CYII=';
